app/requirements.txt
python-telegram-bot
matplotlib
python-telegram-bot
matplotlib
pandas
sentry-sdk
